% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_to_map.R
\name{fb_to_map}
\alias{fb_to_map}
\title{converts fieldbook to field map}
\usage{
fb_to_map(DF, rep = "REP", blk = "BLK", plot = "PLOT", variable = "HI")
}
\arguments{
\item{DF}{a fieldbook data frame}

\item{rep}{column short label used for replication}

\item{blk}{column short label used for block (ignored currently)}

\item{plot}{column short label for plot ID}

\item{variable}{the main variable value to plot}
}
\description{
Simple implementation; currently only works for 1 factorials with equal plots per
replication (as in an RCBD trial).
}

