% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_material_total.R
\name{get_fieldbook_data}
\alias{get_fieldbook_data}
\title{get fieldbook data}
\usage{
get_fieldbook_data(aname)
}
\arguments{
\item{aname}{of a fieldbook (unique)}
}
\description{
get fieldbook data
}

