% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_material_total.R
\name{get_fieldbook_list}
\alias{get_fieldbook_list}
\title{get fieldboook list}
\usage{
get_fieldbook_list(crop = "all", name_only = FALSE, full_path = FALSE)
}
\arguments{
\item{crop}{a crop}

\item{name_only}{only name}

\item{full_path}{full path}
}
\description{
get fieldbook list
}

