% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_material_total.R
\name{get_fieldbook_list_by_loc}
\alias{get_fieldbook_list_by_loc}
\title{get fieldbooks}
\usage{
get_fieldbook_list_by_loc(loc)
}
\arguments{
\item{loc}{location short name}
}
\description{
get fieldbooks
}

