% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_material_total.R
\name{get_fieldbook_total}
\alias{get_fieldbook_total}
\title{get fieldbook total}
\usage{
get_fieldbook_total(crop = "all")
}
\arguments{
\item{crop}{a crop name or 'all' (default)}
}
\description{
get fieldbook total
}
\author{
Reinhard Simon
}

