% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_material_total.R
\name{get_genotype_list_by_loc}
\alias{get_genotype_list_by_loc}
\title{get unique genotypes by location}
\usage{
get_genotype_list_by_loc(loc)
}
\arguments{
\item{loc}{location short name}
}
\description{
(up to hundred)
}

