% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_material_list.R, R/get_material_total.R
\name{get_material_total}
\alias{get_material_total}
\title{get_material_total}
\usage{
get_material_total(crop = "all")

get_material_total(crop = "all")
}
\arguments{
\item{crop}{a crop name or 'all' (default)}
}
\description{
Counts total number of unique plant materials

get materials total.
}
\author{
Reinhard Simon

Reinhard Simon
}

