% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_material_lists.R
\name{get_materials_table}
\alias{get_materials_table}
\title{get_materials_table}
\usage{
get_materials_table(crop, year, mlist_name)
}
\arguments{
\item{crop}{character}

\item{year}{year integer}

\item{mlist_name}{character}
}
\value{
data.frame
}
\description{
always returns a table
}
\author{
Reinhard Simon
}

