% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_material_total.R
\name{get_trial_variables}
\alias{get_trial_variables}
\title{get trial variables}
\usage{
get_trial_variables(trial_name)
}
\arguments{
\item{trial_name}{a trial name}
}
\value{
vector of labels
}
\description{
get trial variables.
}
