% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_material_list.R
\name{import_list_from_prior}
\alias{import_list_from_prior}
\title{import_list_from_prior}
\usage{
import_list_from_prior(crop, year, fname = NULL, year_new,
  mlist_name = NULL, notes = NULL)
}
\arguments{
\item{crop}{character; crop}

\item{year}{integer; source year}

\item{fname}{an Excel file name}

\item{year_new}{integer; new year}

\item{mlist_name}{list; new name}

\item{notes}{character}
}
\value{
boolean TRUE if successful
}
\description{
Imports list from a prior one in the local database if no filename (fname)
is given. Otherwise reads the file (in .xlsx format).
}
\author{
Reinhard Simon
}

