% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_material_lists.R
\name{list_material_lists}
\alias{list_material_lists}
\title{list_material_lists}
\usage{
list_material_lists()
}
\value{
list
}
\description{
lists material-lists
}
\author{
Reinhard Simon
}

