% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_material_list.R
\name{list_years_for_crop}
\alias{list_years_for_crop}
\title{list_years_for_crop}
\usage{
list_years_for_crop(crop)
}
\arguments{
\item{crop}{character}
}
\value{
vector of integer
}
\description{
list years
}
\author{
Reinhard Simon
}
