% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_material_list.R
\name{new_materials_table}
\alias{new_materials_table}
\title{new_materials_table}
\usage{
new_materials_table()
}
\value{
data.frame
}
\description{
creates an empty table with dummy values
}
\author{
Reinhard Simon
}

