% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_material_list.R
\name{server_material_list}
\alias{server_material_list}
\title{server_material_list}
\usage{
server_material_list(input, output, session, dom = "hot_materials", values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{dom}{target dom element name}

\item{values}{reactive values}
}
\description{
Constructs table
}
\author{
Reinhard Simon
}

