% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_material_list.R
\name{ui_material_list}
\alias{ui_material_list}
\title{ui_material_list}
\usage{
ui_material_list(type = "tab", title = "Material list configuration",
  name = "resource_material_list", output = "hot_materials")
}
\arguments{
\item{type}{of ui Element; default is a tab in a shiny dashboard}

\item{title}{display title}

\item{name}{a reference name}

\item{output}{name of output element}
}
\description{
returns a re-usable user interface element
}
\author{
Reinhard Simon
}

