% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_material_list_params.R
\name{ui_material_list_params}
\alias{ui_material_list_params}
\title{UI material list paramters}
\usage{
ui_material_list_params(name = "resource_material_list")
}
\arguments{
\item{name}{character}
}
\description{
An interface to parameters
}
\author{
Reinhard Simon
}
