% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_material_lists.R
\name{get_material_total}
\alias{get_material_total}
\title{get_material_total}
\usage{
get_material_total()
}
\description{
Counts total number of unique plant materials
}
\author{
Reinhard Simon
}

