% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_material_lists.R
\name{import_list_from_prior}
\alias{import_list_from_prior}
\title{import_list_from_prior}
\usage{
import_list_from_prior(crop, year, fname = NULL, mlist_name = NULL)
}
\arguments{
\item{crop}{character}

\item{year}{integer}

\item{fname}{an Excel file name}

\item{mlist_name}{list name}
}
\value{
boolean TRUE if successful
}
\description{
Imports list from a prior one in the local database if no filename (fname)
is given. Otherwise reads the file (in .xlsx format).
}
\author{
Reinhard Simon
}

