% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_material_list.R
\name{list_material_lists}
\alias{list_material_lists}
\title{list_material_lists}
\usage{
list_material_lists(crop = NULL, year = NULL, short = FALSE)
}
\arguments{
\item{crop}{character}

\item{year}{integer}

\item{short}{short list name}
}
\value{
list
}
\description{
lists material-lists
}
\author{
Reinhard Simon
}

