% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_material_lists.R
\name{post_material_table}
\alias{post_material_table}
\title{post_material_table}
\usage{
post_material_table(table_materials, crop, year, mlist_name, notes)
}
\arguments{
\item{table_materials}{a data frame}

\item{crop}{character}

\item{year}{year integer}

\item{mlist_name}{character}

\item{notes}{character}
}
\description{
store program table
}
\author{
Reinhard Simon
}

